Remove mypy ignore annotations
The fms project now includes a (partial) py.typed file, which allows
mypy testing to pass for the fms-extras package.

Typing is incomplete on fms side in the hf package, so we continue to
exclude the corresponding package on fms-extras to avoid failures.

The extra testing uncoverved two issues in calico.py:

- fms_extras/models/calico.py:205: error: "list[int]" has no attribute
  "values"  [attr-defined]

Fixed by removing the redundant values(), set(list[int]) works as it is.

- fms_extras/models/calico.py:217: error: Incompatible types in assignment
  (expression has type Module, variable has type "CalicoBlock")  [assignment]

Fixed by introducing an extra variable instead of overriding the
original variable with an incompatible type.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>